+++ src/Microsoft.ServiceFabric.Client/IApplicationClient_.cs
@@ -1,0 +1,21 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
+namespace Microsoft.ServiceFabric.Client
+{
+    using System;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Microsoft.ServiceFabric.Common;
+    using Microsoft.ServiceFabric.Common.Exceptions;
+    using Microsoft.ServiceFabric.Client.Exceptions;
+
+    /// <summary>
+    /// Interface containing methods for performing ApplicationClient operataions.
+    /// </summary>
+    public partial interface IApplicationClient
+    {
+    }
+}
