+++ src/Microsoft.ServiceFabric.Client/IClusterClient.cs
@@ -711,12 +711,5 @@ namespace Microsoft.ServiceFabric.Client
         Task<AadMetadataObject> GetAadMetadataAsync(
             long? serverTimeout = 60,
             CancellationToken cancellationToken = default(CancellationToken));
-
-
-        /// <summary>
-        /// Gets the connection string for the image store on the current cluster
-        /// </summary>
-        /// <returns></returns>
-        Task<string> GetImageStoreConnectionString();
     }
 }
